refactor: extract LTX2 image encoding from denoising stage into LTX2ImageEncodingStage
